app/requirements.txt
openai
python-dotenv>=1.0.0
fastapi
uvicorn[standard]>=0.29.0
httpx>=0.24.0
black
pytest
typer
twilio
openai-agents
python-multipart
elevenlabs
slowapi
sqlalchemy
psycopg2-binary
langchain
langchain-openai
langchain-community
